Add function `find_multi` to Map and Hashtbl
Map and Hashtbl already contain functions `add_multi` and `remove_multi` to deal
with maps and hashtables when the data type of the values is a list.

`find_multi table key` returns the value associated to `key` if there is some,
or the empty list. `find_multi` raises no exception and
   find_multi : ('key, 'a list) t -> 'key -> 'a list