Fix the generation of mpopcnt.sexp on Windows
On Windows, the echo command will include the quotes passed to it, so
echo '()' results in the literal string `'()'` being written to the
file, which is obviously not valid. Removing the quotes results in
invalid bash as the brackets of course imply a subshell.

The solution is to assume that echo will never gain a `-m` parameter
which means that `echo -mpopcnt` should work correctly on both
platforms. The generation is converted to use `(progn ...)` and add the
brackets from the jbuild file instead of the shell.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>